Bugfix: reveal and checkNeighbors
checkNeighbors: Was missing the top middle and top right spaces, and was instead double counting for those. These are in the last two if statements

reveal: The range was missing the upper bound. extended upper edge of range by 1. Also added a second if statement that prevented endless recursion revealing spaces that were already revealed